refactor(services): add services interfaces
