+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/controller/AuthenticationController.java
@@ -1,9 +1,9 @@
 package com.kopchak.worldoftoys.controller;
 
 import com.kopchak.worldoftoys.dto.TokenAuthDto;
-import com.kopchak.worldoftoys.service.AuthenticationService;
 import com.kopchak.worldoftoys.dto.UserAuthDto;
 import com.kopchak.worldoftoys.dto.UserRegisterDto;
+import com.kopchak.worldoftoys.service.AuthenticationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,17 +16,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/auth")
 @RequiredArgsConstructor
 public class AuthenticationController {
-    private final AuthenticationService service;
+    private final AuthenticationService authenticationService;
 
     @PostMapping("/register")
     public ResponseEntity<TokenAuthDto> register(@RequestBody UserRegisterDto userRegisterDto) {
         return ResponseEntity.status(HttpStatus.CREATED)
-                .body(service.register(userRegisterDto));
+                .body(authenticationService.register(userRegisterDto));
     }
 
     @PostMapping("/authenticate")
     public ResponseEntity<TokenAuthDto> authenticate(@RequestBody UserAuthDto userAuthDto) {
         return ResponseEntity.status(HttpStatus.OK)
-                .body(service.authenticate(userAuthDto));
+                .body(authenticationService.authenticate(userAuthDto));
     }
 }
