+++ worldoftoys/src/main/java/com/kopchak/worldoftoys/filter/JwtAuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.kopchak.worldoftoys.filter;
 
-import com.kopchak.worldoftoys.service.JwtTokenService;
 import com.kopchak.worldoftoys.repository.TokenRepository;
+import com.kopchak.worldoftoys.service.JwtTokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
